fix: random 메소드 수정

+++ backend/src/main/java/life/bareun/diary/product/service/ProductServiceImpl.java
@@ -94,14 +94,14 @@ public class ProductServiceImpl implements ProductService {
             .toList();
 
         // 끝 범위는 전체 색상 가중치의 합
-        double bound = streakColorGrades.stream()
+        float bound = (float) streakColorGrades.stream()
             .mapToDouble(val -> Double.valueOf(val.getWeight()))
             .sum();
 
         // origin 이상 bound 미만의 무작위 실수 값
         // 구간 [1.0, 101.0)의 수 중 랜덤 값을 추출하고
         // 열린 구간 (100.0, 101.0) 사이의 수가 추출되면 100.0으로 치환한다.
-        double gotchaGradeWeight = Math.min(RANDOM.nextDouble(bound) + 1, bound);
+        double gotchaGradeWeight = Math.min(RANDOM.nextFloat(bound) + 1.0, bound);
 
         // 가중치를 반영한 랜덤 값을 뽑기 위한 변수
         double weightSum = 0.0;
